13.2 challenges shop order - update

+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/shop/ComputerOrderRepository.java
@@ -1,10 +1,11 @@
 package com.kodilla.good.patterns.challenges.shop;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public class ComputerOrderRepository  implements OrderRepository{
+public class ComputerOrderRepository implements OrderRepository{
     @Override
-    public void createOrder(User user, double price, LocalDateTime orderDate) {
+    public void createOrder(User user, BigDecimal price, LocalDateTime orderDate) {
         System.out.println("Zamówienie komputera zostało utworzone: Cena: " + price + "; data zamówienia: " + orderDate);
 
     }
