+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/shop/ComputerOrderService.java
@@ -1,10 +1,11 @@
 package com.kodilla.good.patterns.challenges.shop;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ComputerOrderService implements OrderService {
     @Override
-    public boolean order(User user, double price, LocalDateTime orderDate) {
+    public boolean order(User user, BigDecimal price, LocalDateTime orderDate) {
         return true;
     }
 }
