+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/shop/OrderRepository.java
@@ -1,7 +1,8 @@
 package com.kodilla.good.patterns.challenges.shop;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface OrderRepository {
-    void createOrder(User user, double price, LocalDateTime orderDate);
+    void createOrder(User user, BigDecimal price, LocalDateTime orderDate);
 }
