+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/shop/OrderRequest.java
@@ -1,14 +1,15 @@
 package com.kodilla.good.patterns.challenges.shop;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class OrderRequest {
 
     private final User user;
-    private final double price;
+    private final BigDecimal price;
     private final LocalDateTime orderDate;
 
-    public OrderRequest(User user, double price, LocalDateTime orderDate) {
+    public OrderRequest(User user, BigDecimal price, LocalDateTime orderDate) {
         this.user = user;
         this.price = price;
         this.orderDate = orderDate;
@@ -18,7 +19,7 @@ public class OrderRequest {
         return user;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
