+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/shop/OrderService.java
@@ -1,8 +1,9 @@
 package com.kodilla.good.patterns.challenges.shop;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface OrderService {
 
-    boolean order(User user, double price, LocalDateTime orderDate);
+    boolean order(User user, BigDecimal price, LocalDateTime orderDate);
 }
